fix: remove svc only from env with `svc delete --env`
Right now, when a customer runs `svc delete --env e`
copilot will delete the svc from the entire application,
including the ECR repo and the SSM parameter. We only
want to do that if the customer entierly deletes the service.

This change updates svc delete to only remove the env/svc
stack when an environment flag is left in.

Future improvements can be to empty the repo if there
are no more deployed services. I don't think this
is a big deal though.

fixes #1126

+++ internal/pkg/cli/svc_delete.go
@@ -21,9 +21,10 @@ import (
 )
 
 const (
-	svcDeleteNamePrompt       = "Which service would you like to delete?"
-	fmtSvcDeleteConfirmPrompt = "Are you sure you want to delete %s from application %s?"
-	svcDeleteConfirmHelp      = "This will remove the service from all environments, delete the local workspace file, and remove ECR repositories."
+	svcDeleteNamePrompt              = "Which service would you like to delete?"
+	fmtSvcDeleteConfirmPrompt        = "Are you sure you want to delete %s from application %s?"
+	fmtSvcDeleteFromEnvConfirmPrompt = "Are you sure you want to delete %s from environment %s?"
+	svcDeleteConfirmHelp             = "This will remove the service from all environments, delete the local workspace file, and remove ECR repositories."
 )
 
 const (
@@ -116,9 +117,22 @@ func (o *deleteSvcOpts) Ask() error {
 		return nil
 	}
 
+	var deletePrompt string
+	if o.EnvName != "" {
+		// When a customer provides a particular environment,
+		// we'll just delete the service from that environment -
+		// but keep it in the app.
+		deletePrompt = fmt.Sprintf(fmtSvcDeleteFromEnvConfirmPrompt, o.Name, o.EnvName)
+	} else {
+		// When there's no env name passed in, we'll completely
+		// remove the service from the application.
+		deletePrompt = fmt.Sprintf(fmtSvcDeleteConfirmPrompt, o.Name, o.appName)
+	}
+
 	deleteConfirmed, err := o.prompt.Confirm(
-		fmt.Sprintf(fmtSvcDeleteConfirmPrompt, o.Name, o.appName),
+		deletePrompt,
 		svcDeleteConfirmHelp)
+
 	if err != nil {
 		return fmt.Errorf("svc delete confirmation prompt: %w", err)
 	}
@@ -128,7 +142,9 @@ func (o *deleteSvcOpts) Ask() error {
 	return nil
 }
 
-// Execute deletes the application's CloudFormation stack, ECR repository, SSM parameter, and local file.
+// Execute deletes the service's CloudFormation stack.
+// If the service is being removed from the application, Execute will
+// also delete the ECR repository, SSM parameter, and local file.
 func (o *deleteSvcOpts) Execute() error {
 	if err := o.appEnvironments(); err != nil {
 		return err
@@ -137,17 +153,23 @@ func (o *deleteSvcOpts) Execute() error {
 	if err := o.deleteStacks(); err != nil {
 		return err
 	}
-	if err := o.emptyECRRepos(); err != nil {
-		return err
-	}
-	if err := o.removeSvcFromApp(); err != nil {
-		return err
-	}
-	if err := o.deleteSSMParam(); err != nil {
-		return err
+
+	// Only remove the app-level infrastructure if
+	// the service is being completely deleted.
+	if o.needsAppCleanup() {
+		if err := o.emptyECRRepos(); err != nil {
+			return err
+		}
+		if err := o.removeSvcFromApp(); err != nil {
+			return err
+		}
+		if err := o.deleteSSMParam(); err != nil {
+			return err
+		}
+
+		log.Successf("Deleted service %s from application %s.\n", o.Name, o.appName)
 	}
 
-	log.Successf("Deleted service %s from application %s.\n", o.Name, o.appName)
 	return nil
 }
 
@@ -156,6 +178,14 @@ func (o *deleteSvcOpts) validateEnvName() error {
 		return err
 	}
 	return nil
+}
+
+func (o *deleteSvcOpts) needsAppCleanup() bool {
+	// Only remove from a service from the app if
+	// we're removing it from every environment.
+	// If we're just removing the service from one
+	// env, we keep the app configuration.
+	return o.EnvName == ""
 }
 
 func (o *deleteSvcOpts) targetEnv() (*config.Environment, error) {
@@ -308,8 +338,11 @@ func BuildSvcDeleteCmd() *cobra.Command {
 		Use:   "delete",
 		Short: "Deletes a service from an application.",
 		Example: `
-  Delete the "test" service.
+  Delete the "test" service from the application.
   /code $ copilot svc delete --name test
+
+  Delete the "test" service from just the prod environment.
+  /code $ copilot svc delete --name test --env prod
 
   Delete the "test" service without confirmation prompt.
   /code $ copilot svc delete --name test --yes`,
